backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0
